+++ R/cor_test_b.R
@@ -313,8 +313,9 @@ cor_test_b.default = function(x,
 }
 
 
-#' @param formula ADD description!
-#' @param data ADD description!
+#' @param formula A formula of the form \eqn{~ u + v}, where each of \code{u} 
+#' and \code{v} are numeric variables giving the data values for one sample.  
+#' @param data A data.frame containing the variables in \code{formula}.
 #' @rdname cor_test_b
 #' @export
 cor_test_b.formula = function(formula,