+++ tests/testthat/test-prop_test_b.R
@@ -9,6 +9,7 @@ test_that("Test prop_test_b for a single population",{
                   19)
   )
   expect_no_error(fita)
+  expect_no_error(summary(fita))
   expect_s3_class(plot(fita),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -17,6 +18,7 @@ test_that("Test prop_test_b for a single population",{
       prop_test_b(14,
                   n_total = 14 + 19)
   )
+  expect_no_error(summary(fitb))
   expect_equal(fita$results,
                fitb$results)
   
@@ -28,6 +30,7 @@ test_that("Test prop_test_b for a single population",{
                   p = 0.45)
   )
   expect_no_error(fitc)
+  expect_no_error(summary(fitc))
   expect_s3_class(plot(fitc),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -44,11 +47,13 @@ test_that("Test prop_test_b for two populations",{
       prop_test_b(c(14,22),
                   c(19,45))
   )
+  expect_no_error(summary(fita))
   expect_no_error(
     fitb <-
       prop_test_b(c(14,22),
                   n_total = c(14,22) + c(19,45))
   )
+  expect_no_error(summary(fitb))
   expect_equal(fita$results,
                fitb$results)
   
@@ -65,6 +70,7 @@ test_that("Test prop_test_b for two populations",{
                   c(19,45),
                   p = 0.45)
   )
+  expect_no_error(summary(fitc))
   expect_equal(fita$results,
                fitc$results)