+++ tests/testthat/test-t_test_b.R
@@ -13,6 +13,7 @@ test_that("Test t_test_b",{
       t_test_b(rnorm(50))
   )
   expect_no_error(fita)
+  expect_no_error(summary(fita))
   expect_s3_class(plot(fita),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -22,6 +23,7 @@ test_that("Test t_test_b",{
                data = data.frame(outcome = rnorm(50)))
   )
   expect_no_error(fitb)
+  expect_no_error(summary(fitb))
   expect_s3_class(plot(fitb),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -31,6 +33,7 @@ test_that("Test t_test_b",{
                rnorm(15,1))
   )
   expect_no_error(fitc)
+  expect_no_error(summary(fitc))
   expect_s3_class(plot(fitc),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -44,6 +47,7 @@ test_that("Test t_test_b",{
                             asdf = rep(c("a","b"),c(50,15))))
   )
   expect_no_error(fitd)
+  expect_no_error(summary(fitd))
   expect_s3_class(plot(fitd),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))
@@ -54,6 +58,7 @@ test_that("Test t_test_b",{
                paired = TRUE)
   )
   expect_no_error(fite)
+  expect_no_error(summary(fite))
   expect_s3_class(plot(fite),
                   c("patchwork","ggplot2::ggplot","ggplot",
                     "ggplot2::gg","S7_object","gg"))